requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
yfinance==0.2.28
gunicorn==20.1.0
